services/test-service/requirements.txt
azure-identity==1.16.0
azure-servicebus==7.12.0
